requirements.txt
keras
numpy
Pillow
streamlit
tensorflow
